axfs_ng_vfs
Struct DirEntry

pub struct DirEntry<M>(/* private fields */);

A directory entry that can represent either a file or directory

This is the primary abstraction for filesystem nodes in the VFS. It contains both the node implementation and metadata about the node’s location in the filesystem hierarchy.

Implementations

impl<M: RawMutex> DirEntry<M>

pub fn inode(&self) -> u64

Returns the inode number of this entry

pub fn filesystem(&self) -> &dyn FilesystemOps<M>

Returns the filesystem operations for this entry

pub fn update_metadata(&self, update: MetadataUpdate) -> VfsResult<()>

Updates the metadata of this entry

pub fn len(&self) -> VfsResult<u64>

Returns the size of this entry in bytes

pub fn sync(&self, data_only: bool) -> VfsResult<()>

Syncs this entry to persistent storage

pub fn is_empty(&self) -> VfsResult<bool>

Returns whether the entry is empty (has zero length).

impl<M: RawMutex> DirEntry<M>

pub fn new_file( node: FileNode<M>, node_type: NodeType, reference: Reference<M>, ) -> Self

Creates a new file entry

pub fn new_dir( node_fn: impl FnOnce(WeakDirEntry<M>) -> DirNode<M>, reference: Reference<M>, ) -> Self

Creates a new directory entry

pub fn metadata(&self) -> VfsResult<Metadata>

Returns the metadata for this entry, including the correct node type

pub fn downcast<T: Send + Sync + 'static>(&self) -> VfsResult<Arc<T>>

Attempts to downcast the underlying node to a specific type

pub fn downgrade(&self) -> WeakDirEntry<M>

Creates a weak reference to this entry

pub fn key(&self) -> ReferenceKey

Returns a unique key for this entry based on its reference

pub fn node_type(&self) -> NodeType

Returns the node type of this entry

pub fn parent(&self) -> Option<Self>

Returns the parent directory entry, if any

pub fn name(&self) -> &str

Returns the name of this entry

pub fn is_root_of_mount(&self) -> bool

Checks if the entry is a root of a mount point.

pub fn is_ancestor_of(&self, other: &Self) -> VfsResult<bool>

Checks if this entry is an ancestor of another entry

pub fn absolute_path(&self) -> VfsResult<PathBuf>

Returns the absolute path to this entry

pub fn is_file(&self) -> bool

Returns true if this entry represents a file

pub fn is_dir(&self) -> bool

Returns true if this entry represents a directory

pub fn as_file(&self) -> VfsResult<&FileNode<M>>

Returns a reference to the file node, or an error if this is not a file

pub fn as_dir(&self) -> VfsResult<&DirNode<M>>

Returns a reference to the directory node, or an error if this is not a directory

pub fn ptr_eq(&self, other: &Self) -> bool

Checks if this entry points to the same object as another entry

pub fn read_link(&self) -> VfsResult<String>

Reads the target of a symbolic link