axfs_ng_vfs
Struct Mountpoint

pub struct Mountpoint<M> { /* private fields */ }

A mountpoint in the filesystem hierarchy

Represents a point where a filesystem is mounted in the directory tree. Each mountpoint contains a root directory entry and may have child mountpoints.

Implementations

impl<M: RawMutex> Mountpoint<M>

pub fn new( fs: &Filesystem<M>, location_in_parent: Option<Location<M>>, ) -> Arc<Self>

Creates a new mountpoint for the given filesystem

Arguments

fs - The filesystem to mount

location_in_parent - Optional location where this mountpoint is mounted in its parent

pub fn new_root(fs: &Filesystem<M>) -> Arc<Self>

Creates a new root mountpoint (with no parent)

pub fn root_location(self: &Arc<Self>) -> Location<M>

Returns a location pointing to the root of this mountpoint

pub fn location(&self) -> Option<Location<M>>

Returns the location in the parent mountpoint.

pub fn is_root(&self) -> bool

Returns true if this is a root mountpoint (has no parent)

pub fn device(self: &Arc<Self>) -> u64

Returns the device ID for this mountpoint