axfs_ng_vfs
Trait DirNodeOps

pub trait DirNodeOps<M: RawMutex>: NodeOps<M> {
    // Required methods
    fn read_dir(
        &self,
        offset: u64,
        sink: &mut dyn DirEntrySink,
    ) -> VfsResult<usize>;
    fn lookup(&self, name: &str) -> VfsResult<DirEntry<M>>;
    fn create(
        &self,
        name: &str,
        node_type: NodeType,
        permission: NodePermission,
    ) -> VfsResult<DirEntry<M>>;
    fn link(&self, name: &str, node: &DirEntry<M>) -> VfsResult<DirEntry<M>>;
    fn unlink(&self, name: &str) -> VfsResult<()>;
    fn rename(
        &self,
        src_name: &str,
        dst_dir: &DirNode<M>,
        dst_name: &str,
    ) -> VfsResult<()>;
}

Operations specific to directory nodes

This trait extends NodeOps with directory-specific operations like listing entries, creating files, and managing the directory structure.

Required Methods

fn read_dir(&self, offset: u64, sink: &mut dyn DirEntrySink) -> VfsResult<usize>

Reads directory entries.

Returns the number of entries read.

Implementations should ensure that . and .. are present in the result.

fn lookup(&self, name: &str) -> VfsResult<DirEntry<M>>

Lookups a directory entry by name.

fn create( &self, name: &str, node_type: NodeType, permission: NodePermission, ) -> VfsResult<DirEntry<M>>

Creates a directory entry.

fn link(&self, name: &str, node: &DirEntry<M>) -> VfsResult<DirEntry<M>>

Creates a link to a node.

fn unlink(&self, name: &str) -> VfsResult<()>

Unlinks a directory entry by name.

If the entry is a non-empty directory, it should return ENOTEMPTY error.

fn rename( &self, src_name: &str, dst_dir: &DirNode<M>, dst_name: &str, ) -> VfsResult<()>

Renames a directory entry, replacing the original entry (dst) if it already exists.

If src and dst link to the same file, this should do nothing and return Ok(()).

The caller should ensure:

If src is a directory, dst must not exist or be an empty directory.

If src is not a directory, dst must not exist or not be a directory.