axfs_ng_vfs
Trait FileNodeOps

pub trait FileNodeOps<M>: NodeOps<M> {
    // Required methods
    fn read_at(&self, buf: &mut [u8], offset: u64) -> VfsResult<usize>;
    fn write_at(&self, buf: &[u8], offset: u64) -> VfsResult<usize>;
    fn append(&self, buf: &[u8]) -> VfsResult<(usize, u64)>;
    fn set_len(&self, len: u64) -> VfsResult<()>;
    fn set_symlink(&self, target: &str) -> VfsResult<()>;
}

Operations specific to file nodes

This trait extends NodeOps with file-specific operations like reading, writing, and resizing files.

Required Methods

fn read_at(&self, buf: &mut [u8], offset: u64) -> VfsResult<usize>

Reads a number of bytes starting from a given offset.

fn write_at(&self, buf: &[u8], offset: u64) -> VfsResult<usize>

Writes a number of bytes starting from a given offset.

fn append(&self, buf: &[u8]) -> VfsResult<(usize, u64)>

Appends data to the file.

Returns (written, offset) where written is the number of bytes written and offset is the new file size.

fn set_len(&self, len: u64) -> VfsResult<()>

Sets the size of the file.

fn set_symlink(&self, target: &str) -> VfsResult<()>

Sets the file’s symlink target.