axfs_ng_vfs
Trait FilesystemOps

pub trait FilesystemOps<M>: Send + Sync {
    // Required methods
    fn name(&self) -> &str;
    fn root_dir(&self) -> DirEntry<M>;
    fn stat(&self) -> VfsResult<StatFs>;
}

Trait for filesystem operations

This trait defines the core operations that any filesystem implementation must provide to work with the VFS layer.

Required Methods

fn name(&self) -> &str

Returns the name of this filesystem type

fn root_dir(&self) -> DirEntry<M>

Returns the root directory entry for this filesystem

fn stat(&self) -> VfsResult<StatFs>

Returns filesystem statistics