axfs_ng_vfs
Trait NodeOps

pub trait NodeOps<M>: Send + Sync {
    // Required methods
    fn inode(&self) -> u64;
    fn metadata(&self) -> VfsResult<Metadata>;
    fn update_metadata(&self, update: MetadataUpdate) -> VfsResult<()>;
    fn filesystem(&self) -> &dyn FilesystemOps<M>;
    fn sync(&self, data_only: bool) -> VfsResult<()>;
    fn into_any(self: Arc<Self>) -> Arc<dyn Any + Send + Sync>;

    // Provided methods
    fn len(&self) -> VfsResult<u64> { ... }
    fn is_empty(&self) -> VfsResult<bool> { ... }
}

Filesystem node operations

This trait defines the operations that can be performed on any filesystem node, whether it’s a file, directory, or other special node type.

Required Methods

fn inode(&self) -> u64

Gets the inode number of the node.

fn metadata(&self) -> VfsResult<Metadata>

Gets the metadata of the node.

fn update_metadata(&self, update: MetadataUpdate) -> VfsResult<()>

Updates the metadata of the node.

fn filesystem(&self) -> &dyn FilesystemOps<M>

Gets the filesystem

fn sync(&self, data_only: bool) -> VfsResult<()>

Synchronizes the file to disk.

fn into_any(self: Arc<Self>) -> Arc<dyn Any + Send + Sync>

Casts the node to a &dyn core::any::Any.

Provided Methods

fn len(&self) -> VfsResult<u64>

Gets the size of the node.

fn is_empty(&self) -> VfsResult<bool>

Returns whether the node is empty (has zero length).